Move main content in base layout out of vue component
