+++ src/components/FeaturedContentCard.astro
@@ -1,5 +1,5 @@
 ---
-import Card from '../components/Card.astro'
+import Card from './Card.astro'
 const { img, title, description, href } = Astro.props;
 const previewImg = img || '/default-cover.svg';
 ---
